Add files via upload
The mobile app used for the exhibition was developed in a block coding based software called Kodular. The app’s code consists of making a Bluetooth connection through a list peaker and sending different functions such as changing instruments, making sound effects and a set of sequences to the Arduino. It can't be exported in any other way than .apk file, so the rest of the explanation of the code is in the PDF. Also, we started developing a similar app, this time for both Android and iOS devices using a native code developing software based on C# called Xamarin. The file is big, so it's in a Drive folder (link in the .txt file).