Add files via upload
We decided to use a TFT LCD Nextion Display to replace part of the strings that a normal guitar would use. In the PDF you can see the explanation of the editor interface and the code for the display. The HMI is the editor file and the tft file is the one to be uploaded to the display using an SD card.